1. RegisterOnCohort page: 1.1 Moved user registration form into RegisterCohortUser dialog
