+++ src/pages/Admin/Users/RegisterOnCohort/RegisterOnCohort.jsx
@@ -51,20 +51,14 @@ import RegisterOnCohortActions from "./Actions/RegisterOnCohortActions";
 import AuthFormControl from "../../../../components/FormControl/AuthFormControl";
 import FormTextField from "../../../../components/TextField/FormTextField";
 import FormSelect from "../../../../components/Select/FormSelect";
-
-/*
-    ==========================
-    =     AUX VARIABLES      =
-    ==========================
-*/
-const rolesList = ["Admin", "Mentor", "Student"];
+import RegisterCohortUser from "./Actions/RegisterCohortUser";
 
 const RegisterOnCohort = () => {
   /*
-        ==========================
-        =          HOOKS         =
-        ==========================
-    */
+    ==========================
+    =          HOOKS         =
+    ==========================
+  */
   const { cohortId } = useParams();
   const axiosPrivate = useAxiosPrivate();
   const navigate = useNavigate();
@@ -72,36 +66,20 @@ const RegisterOnCohort = () => {
   const { setAuth } = useAuth();
 
   /*
-        ==========================
-        =         STATES         =
-        ==========================
-    */
+    ==========================
+    =         STATES         =
+    ==========================
+  */
   //Fetched data states:
   const [cohortUsers, setCohortUsers] = useState([]);
   const [cohortData, setCohortData] = useState({});
-  //Form states
-  const [userRoles, setUserRoles] = useState([]);
-  const [formError, setFormError] = useState({
-    userNameError: {
-      error: false,
-      errorMessage: "Please enter a valid name",
-    },
-    userEmailError: {
-      error: false,
-      errorMessage: "Please enter a valid e-mail address",
-    },
-    userRolesError: {
-      error: true,
-      errorMessage: "Please select a role for this user",
-    },
-  });
-  const [reset, setReset] = useState(false);
-
+  // Dialog states
+  const [openNewUserDialog, setOpenNewUserDialog] = useState(false);
   /*
-        ==========================
-        =      AUX VARIABLES     =
-        ==========================
-    */
+    ==========================
+    =      AUX VARIABLES     =
+    ==========================
+  */
   const columns = [
     { field: "id", headerName: "ID", minWidth: 100, maxWidth: 130, flex: 1 },
     {
@@ -160,10 +138,10 @@ const RegisterOnCohort = () => {
     },
   ];
   /*
-        ==========================
-        =    ASYNC FUNCTIONS     =
-        ==========================
-    */
+    ==========================
+    =    ASYNC FUNCTIONS     =
+    ==========================
+  */
   const fetchCohortUsers = async () => {
     try {
       const response = await axiosPrivate(`cohort/${cohortId}`);
@@ -211,127 +189,23 @@ const RegisterOnCohort = () => {
   };
 
   /*
-        ==========================
-        =   HANDLER FUNCTIONS    =
-        ==========================
-    */
-
-  // onSelect Role:
-  const handleOnSelectRole = (selectedRoleName) => {
-    setUserRoles(selectedRoleName);
-  };
-
-  // User name:
-  const handleUserNameError = (inputError) => {
-    setFormError((prevState) => ({
-      ...prevState,
-      userNameError: {
-        ...prevState.userNameError,
-        error: inputError,
-      },
-    }));
-  };
-
-  //User email
-  const handleUserEmailError = (inputError) => {
-    setFormError((prevState) => ({
-      ...prevState,
-      userEmailError: {
-        ...prevState.userEmailError,
-        error: inputError,
-      },
-    }));
-  };
-
-  //Form submit:
-  const handleRegisterOnCohortSubmit = async (event) => {
-    event.preventDefault();
-    const formattedUserRegistration = {
-      users: [
-        {
-          name: event.target.userName.value.trim(),
-          email: event.target.userEmail.value.trim(),
-          role: userRoles.map((role) => role.toLowerCase()),
-        },
-      ],
-      cohort: cohortData.cohortId,
-    };
-    const errors = Object.values(formError);
-    try {
-      if (!errors.some((error) => error.error === true)) {
-        const response = await axiosPrivate.post(
-          "auth/register",
-          formattedUserRegistration
-        );
-        if (response.data.users.length > 0) {
-          setCohortUsers((prevState) => [
-            ...prevState,
-            {
-              id: response.data.users[0]._id,
-              userName: response.data.users[0].name,
-              userEmail: response.data.users[0].email,
-              userRole: response.data.users[0].role,
-              userActivatedStatus: response.data.users[0].isActivated,
-            },
-          ]);
-          setReset(true);
-          setUserRoles([]);
-        } else if (response.data.errors.length > 0) {
-          console.error(response.data.errors);
-        }
-      } else {
-        console.error(
-          "There is an error preventing the form submission: check that your entires are correctly validated"
-        );
-      }
-    } catch (error) {
-      if (error.response.status === 403) {
-        console.error(error);
-        //User is required to validate auth again
-        navigate("/login", { state: { from: location }, replace: true });
-        setAuth({
-          userId: "",
-          userName: "",
-          userEmail: "",
-          role: [],
-          loggedIn: false,
-          avatarUrl: "",
-          isActive: undefined,
-          accessToken: "",
-        });
-      } else {
-        console.error(error);
-      }
-    }
-  };
-
-  //. Function to navigate back
+    ==========================
+    =   HANDLER FUNCTIONS    =
+    ==========================
+  */
+  //1. Function to navigate back
   const goBack = () => {
     navigate(-1);
   };
 
   /* 
-        ==========================
-        =        EFFECTS         =
-        ==========================
-    */
+    ==========================
+    =        EFFECTS         =
+    ==========================
+  */
   useEffect(() => {
     fetchCohortUsers();
   }, []);
-
-  useEffect(() => {
-    setFormError((prevState) => ({
-      ...prevState,
-      userRolesError: {
-        ...prevState.userRolesError,
-        error: userRoles.length === 0 ? true : false,
-      },
-    }));
-  }, [userRoles]);
-
-  useEffect(() => {
-    setReset(false);
-  });
 
   return (
     <Container maxWidth="lg">
@@ -371,7 +245,7 @@ const RegisterOnCohort = () => {
               text={"Add new user"}
               type="button"
               width="100%"
-              handlerFunction={() => {}}
+              handlerFunction={() => setOpenNewUserDialog(true)}
             >
               <PersonAddRounded fontSize="large"></PersonAddRounded>
             </AppButton>
@@ -384,7 +258,7 @@ const RegisterOnCohort = () => {
               <PersonSearch fontSize="large"></PersonSearch>
             </AppButton>
             <AppButton
-              text={"Import users from Slack"}
+              text={"Add users from Slack"}
               type="button"
               width="100%"
               handlerFunction={() => {}}
@@ -393,111 +267,6 @@ const RegisterOnCohort = () => {
             </AppButton>
           </AuthFormControl>
         </div>
-        {/*
-        <Box
-          component={"form"}
-          sx={{
-            display: "flex",
-            flexDirection: "column",
-            justifyContent: "center",
-            alignItems: "center",
-            width: "100%",
-          }}
-          autoComplete="off"
-          onSubmit={handleRegisterOnCohortSubmit}
-        >
-          <div className={styles.formContainer}>
-            <AuthFormControl width="75%">
-              <Box
-                sx={{
-                  display: "flex",
-                  flexDirection: "column",
-                  justifyContent: "center",
-                }}
-              >
-                <BadgeRounded fontSize="large"></BadgeRounded>
-                <br></br>
-                <br></br>
-              </Box>
-              <FormTextField
-                required
-                type="text"
-                label="Full name:"
-                name="userName"
-                isFocused={true}
-                width="100%"
-                variant="light"
-                regex={/^[a-zA-z]+([\s][a-zA-Z]+)*$/}
-                onHandleError={handleUserNameError}
-                errorMessage={"Please enter a valid name"}
-                reset={reset}
-              ></FormTextField>
-            </AuthFormControl>
-            <AuthFormControl width="75%">
-              <Box
-                sx={{
-                  display: "flex",
-                  flexDirection: "column",
-                  justifyContent: "center",
-                }}
-              >
-                <Email fontSize="large"></Email>
-                <br></br>
-                <br></br>
-              </Box>
-              <FormTextField
-                required
-                type="text"
-                label="E-mail:"
-                name="userEmail"
-                isFocused={false}
-                width="100%"
-                variant="light"
-                regex={/^[^\s@]+@[^\s@]+\.[^\s@]+$/}
-                onHandleError={handleUserEmailError}
-                errorMessage={"Please enter a valid e-mail address"}
-                reset={reset}
-              ></FormTextField>
-            </AuthFormControl>
-            <AuthFormControl width="75%">
-              <Box
-                sx={{
-                  display: "flex",
-                  flexDirection: "column",
-                  justifyContent: "center",
-                }}
-              >
-                <AdminPanelSettingsRounded fontSize="large" />
-                <br></br>
-              </Box>
-              <AuthFormControl
-                width="100%"
-                isNested={true}
-                error={formError.userRolesError.error}
-              >
-                <FormSelect
-                  id={"userRoles"}
-                  name={"userRoles"}
-                  label={"Roles:"}
-                  selectValue={userRoles}
-                  onSelectValue={handleOnSelectRole}
-                  list={rolesList}
-                  variant={"light"}
-                  multiple={true}
-                  error={formError.userRolesError}
-                ></FormSelect>
-              </AuthFormControl>
-            </AuthFormControl>
-            <AppButton
-              text={"Add new user"}
-              type="submit"
-              width="25%"
-              handlerFunction={() => {}}
-            />
-          </div>
-        </Box>
-      */}
-        
         <AppDataGrid
           columns={columns}
           rows={cohortUsers}
@@ -515,6 +284,7 @@ const RegisterOnCohort = () => {
             }}
           ></AppButton>
         </div>
+        {openNewUserDialog ? <RegisterCohortUser open={openNewUserDialog} handleOpen={setOpenNewUserDialog} onRegisterOnCohortSubmit={setCohortUsers}></RegisterCohortUser> : null}
       </Paper>
     </Container>
   );
